requirements.txt
appdirs==1.4.3
asgiref==1.1.2
attrs==17.2.0
autobahn==17.5.1
Automat==0.6.0
channels==1.1.3
constantly==15.1.0
daphne==1.2.0
Django==1.11.1
django-countries==4.5
django-haystack==2.6.1
django-languages==0.1
djangorestframework==3.6.3
incremental==17.5.0
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
Twisted==17.1.0
txaio==2.7.1
zope.interface==4.4.1
